chore: Fix Z-index type

+++ src/Uno.UI/UI/Xaml/Controls/Canvas/Canvas.skia.cs
@@ -2,7 +2,7 @@
 
 public partial class Canvas
 {
-	static partial void OnZIndexChangedPartial(UIElement element, double? zindex)
+	static partial void OnZIndexChangedPartial(UIElement element, int? zindex)
 	{
 		element.Visual.ZIndex = (int)zindex;
 	}
