+++ src/Uno.UI/UI/Xaml/Controls/Canvas/Canvas.wasm.cs
@@ -2,7 +2,7 @@ namespace Windows.UI.Xaml.Controls
 {
 	partial class Canvas
 	{
-		static partial void OnZIndexChangedPartial(UIElement element, double? zindex)
+		static partial void OnZIndexChangedPartial(UIElement element, int? zindex)
 		{
 			if (zindex is { } d)
 			{
